+++ src/main/resources/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
+## 1.0.2 (2023-10-27)
+- `Main` : Call to get recombination fraction between 2 positions
+- `MathLib` : Externalisation of Math calls
+- `Unphased` : Process a new data format for unphased data
+- `InputFile` : Fix export to estiage format hen mixing snp/indels
 ## 1.0.1 (2022-08-01)
-* `Estiage Format`: **optimized** Check Odd/Even Microsat
-* `Estiage Format`: **added** Conversion from pre-estiage format to estiage (200->202 to 1->2)
+- `Estiage Format`: **optimized** Check Odd/Even Microsat
+- `Estiage Format`: **added** Conversion from pre-estiage format to estiage (200->202 to 1->2)